backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.6
sqlparse==0.4.3
zope.event==4.5.0
zope.interface==5.4.0
